selenium-s3-screenshot/requirements.txt
attrs==25.3.0
boto3==1.38.37
boto3-stubs==1.38.37
botocore==1.38.37
certifi==2025.6.15
charset-normalizer==3.4.2
h11==0.16.0
idna==3.10
jmespath==1.0.1
outcome==1.3.0.post0
packaging==25.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.4
s3transfer==0.13.0
selenium==4.33.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
urllib3==2.4.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
